ref: Move Endpoint.channelShims to EndpointShim.
